Declare time and file libc functions with their C99/POSIX widths
The bundled headers declared several libc functions whose real return
or parameter types are wider than `int`. The codegen narrows a libc
return to the declared type and narrows an argument to the declared
parameter type, so an `int` declaration truncated 64-bit values:

  time.h   - time / clock / mktime return time_t / clock_t (64-bit);
             difftime takes two time_t by value; localtime, gmtime,
             gmtime_r, localtime_r, ctime_r and time take a time_t*.
             clock() overflowed `int` after ~35 min of CPU time
             (CLOCKS_PER_SEC == 1e6); difftime truncated arguments
             past 2^31 before the subtraction.
  unistd.h - lseek returns off_t and takes an off_t offset; ftruncate
             takes an off_t length; sysconf / pathconf return long.
  stdio.h  - fseek takes a long offset; ftell returns long.

Declare each with its C99 7.19 / 7.23 / POSIX type. `long` matches
off_t / time_t / clock_t on LP64; on LLP64 `time_t` is already a
typedef to `long long`. The spurious post-call narrowing drops out
(visible in the sysconf_pagesize and stdio_fgetpos_roundtrip
snapshots). The time_t* pointer parameters are codegen-neutral.

Regression: libc_time_widths.c checks difftime with arguments past
2^32 and time()'s return / out-parameter agreement.

+++ headers/include/unistd.h
@@ -196,9 +196,12 @@ long pread64(int fd, void *buf, unsigned long n, long offset);
 long pwrite64(int fd, const void *buf, unsigned long n, long offset);
 #endif
 int access(char *path, int mode);
-int lseek(int fd, int offset, int whence);
+// POSIX: lseek returns off_t and takes an off_t offset; ftruncate takes an
+// off_t length. off_t is 64-bit, so `int` truncates offsets/lengths past
+// 2GB. `long` matches off_t on LP64 (the POSIX targets this block serves).
+long lseek(int fd, long offset, int whence);
 int fsync(int fd);
-int ftruncate(int fd, int len);
+int ftruncate(int fd, long len);
 int fcntl(int fd, int cmd, ...);
 int stat(char *path, char *buf);
 int lstat(char *path, char *buf);
@@ -241,8 +244,9 @@ int chown(char *path, int uid, int gid);
 int truncate(char *path, int len);
 int link(char *from, char *to);
 int symlink(char *from, char *to);
-int pathconf(char *path, int name);
-int sysconf(int name);
+// POSIX: pathconf / sysconf return long; some limits exceed 32 bits.
+long pathconf(char *path, int name);
+long sysconf(int name);
 int getrusage(int who, char *usage);
 int flock(int fd, int operation);
 int nanosleep(char *req, char *rem);
